initial llvm16 rework
