+++ packages/cras.rb
@@ -28,7 +28,7 @@ class Cras < Package
   depends_on 'gtest' => :build
   depends_on 'iniparser' # R
   depends_on 'ladspa'
-  depends_on 'llvm_dev16' => :build
+  depends_on 'llvm16_dev' => :build
   depends_on 'rust' => :build
   depends_on 'sbc' # R
   depends_on 'speexdsp' # R
