+++ packages/crystal.rb
@@ -9,7 +9,7 @@ class Crystal < Package
   source_url 'https://github.com/crystal-lang/crystal/releases/download/1.6.0/crystal-1.6.0-1-linux-x86_64-bundled.tar.gz'
   source_sha256 '5b1b287563839c3c9a2fa7b620c48b5c56de0756464e8370f3a7f720a3ba4df0'
 
-  depends_on 'llvm_lib16' # R
+  depends_on 'llvm16_lib' # R
 
   def self.install
     FileUtils.mkdir_p CREW_DEST_PREFIX.to_s
