+++ packages/inkscape.rb
@@ -59,7 +59,7 @@ class Inkscape < Package
   depends_on 'libxext' # R
   depends_on 'libxml2' # R
   depends_on 'libxslt' # R
-  depends_on 'llvm_lib16' => :build
+  depends_on 'llvm16_lib' => :build
   depends_on 'pangomm_1_4' # R
   depends_on 'pango' # R
   depends_on 'poppler' # R
