+++ packages/js102.rb
@@ -27,7 +27,7 @@ class Js102 < Package
   depends_on 'glibc' # R
   depends_on 'icu4c' # R
   depends_on 'libnotify' => :build
-  depends_on 'llvm_dev16' => :build
+  depends_on 'llvm16_dev' => :build
   depends_on 'ncurses' # R
   depends_on 'nspr'
   depends_on 'nss' # R
