+++ packages/js78.rb
@@ -24,7 +24,7 @@ class Js78 < Package
 
   depends_on 'autoconf213' => :build
   depends_on 'rust' => :build
-  depends_on 'llvm_dev16' => :build
+  depends_on 'llvm16_dev' => :build
   depends_on 'nspr'
 
   @rust_default_host = case ARCH
